Diode Eqn Fitting
Was trying to see if I could add a series resistance term to the diode equation. This was not proving very successful. I'm getting a fit in certain cases but the parameters don't seem to make physical sense. Strange things are happening when I try and vary the temperature. Maybe my fit model is too complicated. Am I overfitting? Using too many parameters?