Diode Eqn Fit
It's still not working properly. The Is parameters is wildly out of scale with the others which means that for the fit model derivative wrt Is is either way too large, or way too small in comparison to other derivatives. Will go with simpler solution solution of fitting to linear portion of graph.