Diode Eqn Fitting
It turns out that I was right the first time. I was using the incorrect diode eqn function when computing the synthetic data and the residuals, but using the correct eqn when computing the fit, so I was fitting the wrong model to the wrong data. Code seems to be working fine now. Best practice is to keep T constant during fit, this gives more realistic values for eta and Is. Fitted values for Rs seem reasonable.